Move DefaultCredentials to own file
